Move SYCL_DeepCopy.hpp into core/src/SYCL
