+++ core/src/SYCL/Kokkos_SYCL_Space.cpp
@@ -42,13 +42,13 @@
 //@HEADER
 */
 
-#include <Kokkos_SYCL_Space.hpp>
-#include <Kokkos_SYCL_DeepCopy.hpp>
 #include <Kokkos_HostSpace.hpp>
-#include <impl/Kokkos_Profiling.hpp>
 #include <Kokkos_SYCL.hpp>
+#include <Kokkos_SYCL_Space.hpp>
+#include <SYCL/Kokkos_SYCL_DeepCopy.hpp>
 #include <SYCL/Kokkos_SYCL_Instance.hpp>
 #include <impl/Kokkos_MemorySpace.hpp>
+#include <impl/Kokkos_Profiling.hpp>
 
 /*--------------------------------------------------------------------------*/
 /*--------------------------------------------------------------------------*/
