+++ core/src/decl/Kokkos_Declare_SYCL.hpp
@@ -47,7 +47,7 @@
 
 #if defined(KOKKOS_ENABLE_SYCL)
 #include <Kokkos_SYCL.hpp>
-#include <Kokkos_SYCL_DeepCopy.hpp>
+#include <SYCL/Kokkos_SYCL_DeepCopy.hpp>
 #include <SYCL/Kokkos_SYCL_Parallel_Range.hpp>
 #include <SYCL/Kokkos_SYCL_Parallel_Reduce.hpp>
 #endif
